+++ CHANGELOG.md
@@ -16,6 +16,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### ✏️ Changed
+
+- Minor improvement in the logger to avoid print empty options object
+
 ## [0.3.0] - 2018-6-27
 
 ### 🐛 Fixed